Thanks to Brandon Craig Rhodes <brandon@oit.gatech.edu>:
While debugging an application that uses PostgreSQL, valgrind (which I
had not used before but am very impressed with) helpfully pointed out
that PostgreSQL often used memory that had already been free()'d.

This turned out to be because the pg_query function, when processing
the query result, calls PQoidValue() on the variable `result' *after*
having just freed the result with a PQclear().  To correct this, the
oid call can simply be moved a few lines earlier: