Change internal names to have leading underscores but not trailing.  This matches Python's method of keeping attributes private.
Comment out the debugging SELECT.  This is due to the fact that argv is
not always available.  It fails in mod_python e.g.  This way it can be
added if so desired.

Wrap close so that we can track the state.

Create new method reopen().  This is useful if we need to reopen the
database without having to track the opening args ourself.  This is
useful in web pages that fork off to do work after the page is delivered.
We would close the db, fork and call reopen.  It is also good for getting
a separate connection if we need to create transaction blocks withing an
existing transaction block.