From: Sergio Lenzi <enigma@lztech.com.br>
Hello,

I am using your excelent python module pygresql for
postgres databases on python and I notice that for the
more recent version of postgres, starting with 7.3.4, it
uses a program pg_config.

I make some "fixes"  in the setup.py code to make use
of the pg_config program, that puts all the libraries and includes
files automaticaly in the setup.py

Enclosed in this email is the diff against your code.

Free to use...

Thanks for the attention,

Sergio