Fixed a problem on NetBSD
It seems isdigit() and ispunct() expect unsigned chars instead of
ints on certain platforms, and silencing them by casting signed
chars to ints was counter productive, causing serious issues on
NetBSD and maybe other platforms. To avoid problems with the
implementation of isdigit() and ispunct() on different platforms
and with different locales, we totally avoid these functions now
and check the chars "manually" instead, which is an easy thing.