Clean up the docs
We had used .txt for the old docs and .rst for the new Sphinx docs,
which unfortunately got out of sync. I have merged the duplicates,
keeping the latest version of the text, and left them only as .txt
files for the time being until we give Sphinx another go.

The index file is not auto generated - the index.rst file is only
a reminder that we want to move everything to Sphinx.