Specify supported Python versions for the 4.x branch
Tests for Python 2.4, 2.5, 2.6., 2.7 are running now, so we can specify these
versions as supported in this branch. The trunk will only support Python >= 2.6.
The test are not compatible with Python 2.3, but the basic functionality still
works in Python 2.3, so old code to support Python < 2.4 has been left in place.

A config file for testing with tox has been added. Unfortunately, tox supports
only Python 2.6 and 2.7. So Python 2.4 and 2.5 need to be tested manually.