Add cursor classes for returning various row types
Currently, PyGreSQL returns rows as lists in the DB-API 2 module.
It is more common to return tuples (and consistent with how it
is done int the classic module). Even better are named tuples
which are cheap and available in all Py versions PyGreSQL supports.

So I made tuples the default now and added additional Cursor types
for returning rows as lists, named tuples, dicts and ordered dicts.
Also added an attribute for setting the default Cursor type in the
connection and special methods for creating the various cursor types
from the same connection.

However, I think this should be simplified and named tuples should
become the default. Named tuples can be easily converted to ordinary
lists and tuples (although this should never be necessary), and they
can also be easily converted to ordered dicts by calling ._asdict().
So I'm planning to change this before long, but for reference I have
checked in this implementation with all necessary tests anyway.