Implement the callproc() cursor method
The implementation has been a bit simplified in that output and
input/output parameters are not changed in the return value, i.e.
they can only be retrieved using the fetch methods. To implement
this, it would be necessary to determine (through a query to the
database catalog) which parameters are output parameters and
fetch the return values for these, which would be very costly.