Assume select privilege in insert/update()
Don't query the database whether we have select privilege on the
table on every call of insert/update(). We assume that we can
update/insert() anyway, and it's pretty sane then to assume we can
select, too. This spares one database request per call that is
superfluous under all normal circumstances. In the theoretically
possible but rare case that you can insert/update and not select,
a normal query() is always possible instead.