Build source distribution with setup script
Instead of building the source distribution (the tarball) manually
in mktar, we now let setup.py do the work for us.

This has the advantage that we can be sure that the created tarball
conforms to the standards (e.g. contains the egg-info bits), and we
don't need to maintain the list of source files in two places any
more - we only need to maintain the MANIFEST.in file now.

Note that setup.py creates a tar.gz file instead of .tgz as before.
We should also distribute it like that because it's the standard on
PyPI etc. I have also changed the download docs to reflect this.