Add more safety checks to the mktar script
Make sure the right files are packaged, and with the right permissions.

Also specified the exact top-level source files in MANIFEST.in,
in order to make sure that other files used for testing etc.
do not accidentally get packaged. The mktar script now also checks
the number of top-level files and warns if something does not fit.