Add typecasting of dates, times, timestamps, intervals
So far, PyGreSQL has returned these types only as strings (in various
formats depending on the DateStyle setting) and left it to the user
to parse and interpret the strings.  These types are now properly cast
into the corresponding detetime types of Python, and this works with
any setting of DatesStyle, even if you change DateStyle in the middle
of a database session.

To implement this, a fast method for getting the datestyle (cached and
without roundtrip to the database) has been added. Also, the typecast
mechanism has been extended so that typecast functions can optionally
also take the connection as argument.

The date and time typecast functions have been implemented in Python
using the new typecast registry and added to both pg and pgdb. Some
duplication of code in the two modules was unavoidable, since we don't
want the modules to be dependent of each other or install additional
helper modules. One day we might want to change this, put everything
in one package and factor out some of the functionality.