Sanitize parsing of kwargs to handle quote and backslash characters
The escaping of quotes in kwargs was done incorrectly, leading to
problems where kwargs or dbname included a single quote.

This caused an exception, pg.InternalError if a database name like
`funny copy"db'with\\quotes` was used:

`missing "=" after "with\\\\quotes'" in connection info string`

But only if kwargs were specified - otherwise, the dbname was passed
safely to _connect() as a plain string without being modified.

This commit addressed this problem by correcting the order in which
escaping is done.

The kwargs test is modified to ensure that the case of included
escape characters and single quotes are covered.

Co-authored-by: Tyler Ramer <tramer@pivotal.io>
Co-authored-by: Jamie McAtamney <jmcatamney@pivotal.io>