Do not close large objects when deallocating (#30)
If a large object goes out of scope and is deallocated after a transaction
is closed, it will have been already closed on the server side and you will
get an error message on the server when the dealloc method tries to close it
again. In other cases, the object might be closed too early when the Python
object falls out of scope. Therefore it is best to not close it here.